viewer: extract DashboardExplorerRedirect out of LocalRouter
The cloud app (core) mounts `/workspace/dashboard/:name/explorer` too, and after
Explore 2.0 that route is this component. It could only be imported from
`LocalRouter.jsx`, which is not importable from core:

    export const localURLConfig = createURLConfig({ environment: 'server' });
    setGlobalURLConfig(localURLConfig);          // module scope
    const LocalRouter = createBrowserRouter(...) // module scope

Importing it would have silently repointed every API call in the cloud app at
the local-serve URL config and stood up a second browser router. The build
would have succeeded — this only shows up at runtime — so it is worth the
extraction rather than a comment telling the next person not to.

Now its own module, re-exported from `LocalRouter` so existing callers are
unaffected. `LocalRouter.test.jsx` passes unmodified, which is the check that
the re-export preserved the contract. The hooks it was the sole user of
(useEffect/useRef/useState/useStore) come out of LocalRouter with it.
